backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
psutil
requests
pydantic
emergentintegrations
